Split GMP runtime helpers and only link libgmp when needed
